+++ templates/esm/routes/api/users.js
@@ -1,19 +1,18 @@
 import * as expressExtension from 'express-sweet';
 import {Router} from 'express';
-import {query, body, validationResult} from 'express-validator';
+import {query, body} from 'express-validator';
 import UserModel from '../../models/UserModel.js';
 import NotFoundError from '../../errors/NotFoundError.js';
-import * as CustomValidation from '../../shared/CustomValidation.js';
+import isValidImageDataUrl from '../../validators/isValidImageDataUrl.js';
+import checkValidationResult from '../../middlewares/checkValidationResult.js';
 
 const router = Router();
 
 router.post('/login', [
   body('email').trim().not().isEmpty().isEmail(),
-  body('password').trim().not().isEmpty()
+  body('password').trim().not().isEmpty(),
+  checkValidationResult,
 ], async (req, res, next) => {
-  const result = validationResult(req);
-  if (!result.isEmpty())
-    return void res.status(400).json({errors: result.array()});
   const isAuthenticated = await expressExtension.services.Authentication.authenticate(req, res, next);
   res.json(isAuthenticated);
 });
@@ -29,11 +28,9 @@ router.get('/', [
   query('length').not().isEmpty().isInt({min: 1}),
   query('order').not().isEmpty().isIn(['name', 'email', 'modified']),
   query('dir').optional({nullable: true, checkFalsy: true}).isIn(['asc', 'desc']),
-  query('search').trim().optional({nullable: true, checkFalsy: true}).trim()
+  query('search').trim().optional({nullable: true, checkFalsy: true}).trim(),
+  checkValidationResult,
 ], async (req, res) => {
-  const result = validationResult(req);
-  if (!result.isEmpty())
-    return void res.status(400).json({errors: result.array()});
   const data = await UserModel.paginate(req.query);
   data.draw = req.query.draw;
   res.json(data);
@@ -41,11 +38,9 @@ router.get('/', [
 
 router.get('/email-exists', [
   query('user.email').trim().not().isEmpty(),
-  query('excludeUserId').optional({nullable: true, checkFalsy: true}).isInt({min: 1})
+  query('excludeUserId').optional({nullable: true, checkFalsy: true}).isInt({min: 1}),
+  checkValidationResult,
 ], async (req, res) => {
-  const result = validationResult(req);
-  if (!result.isEmpty())
-    return void res.status(400).json({errors: result.array()});
   const emailExists = await UserModel.emailExists(req.query.user.email, req.query.excludeUserId || null);
   res.json({valid: !emailExists});
 }); 
@@ -54,11 +49,9 @@ router.post('/', [
   body('user.email').trim().not().isEmpty().isEmail(),
   body('user.name').trim().not().isEmpty().isLength({max: 30}),
   body('user.password').trim().not().isEmpty().isLength({max: 128}),
-  body('user.icon').not().isEmpty().custom(CustomValidation.isImageDataUrl)
+  body('user.icon').not().isEmpty().custom(isValidImageDataUrl),
+  checkValidationResult,
 ], async (req, res) => {
-  const result = validationResult(req);
-  if (!result.isEmpty())
-    return void res.status(400).json({errors: result.array()});
   await UserModel.createUser(req.body.user);
   res.json(true);
 });
@@ -72,12 +65,10 @@ router.put('/:userId(\\d+)', [
   body('user.email').trim().not().isEmpty().isEmail(),
   body('user.name').trim().not().isEmpty().isLength({max: 30}),
   body('user.password').trim().optional({nullable: true, checkFalsy: true}).isLength({max: 128}),
-  body('user.icon').not().isEmpty().custom(CustomValidation.isImageDataUrl)
+  body('user.icon').not().isEmpty().custom(isValidImageDataUrl),
+  checkValidationResult,
 ], async (req, res, next) => {
   try {
-    const result = validationResult(req);
-    if (!result.isEmpty())
-      return void res.status(400).json({errors: result.array()});
     await UserModel.updateUser(req.params.userId, req.body.user);
     res.json(true);
   } catch (error) {
@@ -97,12 +88,10 @@ router.put('/profile', [
   body('user.email').trim().not().isEmpty().isEmail(),
   body('user.name').trim().not().isEmpty().isLength({max: 30}),
   body('user.password').trim().optional({nullable: true, checkFalsy: true}).isLength({max: 128}),
-  body('user.icon').not().isEmpty().custom(CustomValidation.isImageDataUrl)
+  body('user.icon').not().isEmpty().custom(isValidImageDataUrl),
+  checkValidationResult,
 ], async (req, res, next) => {
   try {
-    const result = validationResult(req);
-    if (!result.isEmpty())
-      return void res.status(400).json({errors: result.array()});
     await UserModel.updateUser(req.user.id, req.body.user);
     res.json(true);
   } catch (error) {