+++ templates/esm/shared/empty.js
@@ -1,8 +1,0 @@
-/**
- * Returns true if the string is empty, false otherwise.
- */
-export default str => {
-  if (str == null || str === false)
-    return true;
-  return str.toString().replace(/^[\s　]+|[\s　]+$/g, '') === '';
-}